+++ src/condor_ckpt_server/fileinfo2.C
@@ -167,7 +167,8 @@ void FileInformation::PrintFileInfo()
   if (count != num_files)
     {
 	  dprintf(D_ALWAYS, "WARNING incorrect count for the number of files "
-			  "(count = %d, num = %d)\n", count, num_files);
+			  "(count = %lu, num = %lu)\n", 
+			  (unsigned long) count, (unsigned long) num_files);
     }
 }
